+++ language/english.php
@@ -1,5 +1,4 @@
 <?php
-//  $Id: english.php 42 2012-05-23 15:05:36Z root $
 /**
 *   Default English Language file for the lgLib plugin.
 *
@@ -56,6 +55,7 @@ $LANG_confignames['lglib'] = array(
     'img_cache_interval' => 'Image Cache Cleaning Interval (minutes)',
     'img_cache_maxage' => 'Cached Image Max Age (days)',
     'slimbox_autoactivation' => 'Auto-Activate Slimbox?',
+    'use_lglib_messages' => 'Use lglib_messages template var?',
 );
 
 $LANG_configsubgroups['lglib'] = array(