+++ language/spanish_colombia_utf-8.php
@@ -55,6 +55,8 @@ $LANG_confignames['lglib'] = array(
     'cron_key' => 'Scheduled Task Security Key',
     'img_cache_interval' => 'Image Cache Cleaning Interval (minutes)',
     'img_cache_maxage' => 'Cached Image Max Age (days)',
+    'use_lglib_messages' => 'Use lglib_messages template var?',
+    'slimbox_autoactivation' => 'Auto-Activate Slimbox?',
 );
 
 $LANG_configsubgroups['lglib'] = array(