+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 lgLib utility plugin for glFusion
 ====================
+1.0.1
+=====
+- Add option to use lglib_messages template variable for messages. Default
+is to append messages to info_block in header.thtml
+
 1.0.0
 =====
 - Add "/index.php" to backup list form URL, delete checkboxes weren't working